Issue #10: Some preparations.
